Filter table A to perform a fair comparison with table B

Filter table A to perform a fair comparison with table B using row.names

comparison(
  tabA,
  tabB,
  use_names,
  summaryze = FALSE,
  formate = TRUE,
  verbose = FALSE
)

Arguments

tabA

Table A, from evaluate (output values)

tabB

Table B, from evaluate (for row.names only)

use_names

list of names to be used instead of row.names of Table B

summaryze

logical, true to update or add a summary

formate

argument passed to evaluation function, only works with summaryze = TRUE

verbose

display additional information

Value

Table A containing only the row.names in table B

Note

to perform a comparison check both tables using comparison(A,B) and comparison(B,A)

Examples

model <- 1:10
data  <- model + rnorm(10,0.2)
A     <- stats(mo = model, ob = data)
for(i in 2:10){
   data  <- model + rnorm(10,0.2)
   A     <- rbind(A,stats(mo = model, ob = data))
}
row.names(A) <- c('a0','a1','a3','b0','b1','b2','b3','d0','d1','d2')

data   <- model + rnorm(10,0.2)
B      <- stats(mo = model, ob = data)
for(i in 2:5){
    data   <- model + rnorm(10,0.2)
    B      <- rbind(B,stats(mo = model, ob = data))

}
row.names(B) <- c('a0','a1','z0','z1','z2')

comparison(A,B,verbose = TRUE)
#> table A:
#>  [1] "a0" "a1" "a3" "b0" "b1" "b2" "b3" "d0" "d1" "d2"
#> table B:
#> [1] "a0" "a1" "z0" "z1" "z2"
#>     n      Obs Sim         r       IOA FA2      RMSE          MB        ME
#> a0 10 5.539969 5.5 0.9360283 0.9631389 0.9 1.1563195 -0.03996905 0.9641047
#> a1 10 5.864619 5.5 0.9758272 0.9827907 0.9 0.7779044 -0.36461853 0.6687720
#>           GE   MFB (%)  MFE (%)    NMB (%)  NME (%)
#> a0 0.9641047  8.309950 25.96380 -0.7214671 17.40271
#> a1 0.6687720 -5.980944 19.56155 -6.2172592 11.40350
comparison(B,A,verbose = TRUE)
#> table A:
#> [1] "a0" "a1" "z0" "z1" "z2"
#> table B:
#>  [1] "a0" "a1" "a3" "b0" "b1" "b2" "b3" "d0" "d1" "d2"
#>     n      Obs Sim         r       IOA FA2      RMSE         MB        ME
#> a0 10 5.667803 5.5 0.8965028 0.9436436 0.8 1.4132247 -0.1678026 1.1152429
#> a1 10 5.373296 5.5 0.9726493 0.9819011 0.9 0.8194234  0.1267040 0.6123097
#>           GE  MFB (%)  MFE (%)   NMB (%)  NME (%)
#> a0 1.1152429 19.86392 45.42526 -2.960629 19.67681
#> a1 0.6123097 15.68749 24.57379  2.358031 11.39542