Creata a NetCDF file with the surface maximum

Read and calculate the maximum value of a variable from a list of wrf output files.

extract_max(
  filelist,
  variable = "o3",
  field = "4d",
  prefix = "max",
  units = "ppmv",
  meta = T,
  filename,
  verbose = TRUE
)

Arguments

filelist

list of files to be read

variable

variable name

field

'4d' (default), '3d', '2d' or '2dz' see notes

prefix

to output file, defolt is serie

units

units on netcdf file (default is ppmv)

meta

use Times, XLONG and XLAT data (only work with 2d variable for file)

filename

name for the file, in this case prefix is not used

verbose

display additional information

Note

The field argument '4d' / '2dz' is used to read a 4d/3d variable droping the 3rd dimention (z).

Examples

dir.create(file.path(tempdir(), "MAX"))
folder <- system.file("extdata",package="hackWRF")
wrf_file <- paste0(folder,"/wrf.day1.o3.nc")
extract_max(filelist = wrf_file,prefix = paste0(file.path(tempdir(),"MAX"),'/mean'))
#> extracting max of o3 field 4d 
#> reading: C:/Users/schuch/AppData/Local/Temp/RtmpKU9eYr/temp_libpath4b5c4183409d/hackWRF/extdata/wrf.day1.o3.nc file 1 of 1 
#> min: 2.123069e-05 mean: 0.007767247 max: 0.03175572 
#> total times: 1 
#> output: C:\Users\schuch\AppData\Local\Temp\Rtmpm0jfzn/MAX/mean.o3.nc