Convert a raster or brick to a array for WRF-Chem

Conversion a raster or brick to a array for WRF-Chem model

raster_to_ncdf(r, na_value = 0)

Arguments

r

input raster

na_value

value for missing values

Value

array

Examples

r <- raster::raster(paste0(system.file("extdata", package = "hackWRF"),
                           "/wrf.day1.o3.nc"))
a <- raster_to_ncdf(r)